разделяет генерацию данных на модули
